requirements.txt
matplotlib
pandas
numpy
py
mysql-connector-python